requirements.txt
gitpython
mcp
neo4j
psycopg
python-dotenv
qdrant-client
tiktoken
